Move Runner methods to Lambda Util
