+++ packages/appsync-emulator-serverless/lambdaRunnerPython.js
@@ -1,27 +1,4 @@
-const log = require('logdown')('appsync-emulator:lambdaRunner');
-
-const parseErrorStack = error =>
-  error.stack
-    .replace(/at /g, '')
-    .split('\n    ')
-    .slice(1);
-
-const sendOutput = output => {
-  process.send({ type: 'success', output }, process.exit);
-};
-const sendErr = err => {
-  let error;
-  if (err instanceof Error) {
-    error = {
-      stackTrace: parseErrorStack(err),
-      errorType: err.constructor.name,
-      errorMessage: err.message,
-    };
-  } else {
-    error = err;
-  }
-  process.send({ type: 'error', error }, process.exit);
-};
+const { log, sendErr, sendOutput, installExceptionHandlers } = require('./lambda/util');
 
 process.once(
   'message',
@@ -29,10 +6,10 @@ process.once(
     try {
         log.info('load', module);
    
-        var spawn = require('child_process').spawn          
-        var args = ["invoke", "local", "-f", handlerMethod]
+        let spawn = require('child_process').spawn;
+        let args = ["invoke", "local", "-f", handlerMethod];
 
-        var sls = spawn("sls", 
+        let sls = spawn("sls",
             args, 
             {
                 env: process.env,
@@ -44,11 +21,11 @@ process.once(
         sls.stdin.write(JSON.stringify(payload) + "\n");
         sls.stdin.end();
         
-        let results = ''
+        let results = '';
         sls.stdout.on('data', function(data) {
-            results = data.toString('utf8');
-            results = results.replace('\n', '')
-        })
+            results = data.toString();
+            results = results.replace('\n', '');
+        });
 
         sls.on('close', function(code) {
             sendOutput(JSON.parse(results));
@@ -59,11 +36,4 @@ process.once(
   },
 );
 
-process.on('uncaughtException', err => {
-  log.error('uncaughtException in lambda', err);
-  process.exit(1);
-});
-process.on('unhandledRejection', err => {
-  log.error('unhandledRejection in lambda', err);
-  process.exit(1);
-});
+installExceptionHandlers();
